Finally kinda got logging done
Sorry that I've been away for so long and have gotten so little
done—I've been busy with things outside of this project.

Logging to a file (both for general logging and for crash logging) still
doesn't work.  I plan on doing some refactoring soon, ideally when I'm
in a better state of mind.