Added a test of debug builds
For some reason, I need to give functions that I want to be overloaded
with a dummy no-op function the `[[clang::overloadable]]` attribute,
but…only in the header files, not the source files.  This causes the
function to be overloaded by the dummy no-op, which is the opposite of
what I would expect to happen.  ??????  Oh well, at least now I know
what to do.