Bronked the makefile to force in a debug target
The way I managed to shimmy in a debug target into this _completely_
misses, like, the _entire point_ of makefiles.  I did try to put some
shell scripting into the targets so that there could be at least some of
the benefits of makefiles (mostly not compiling redundant object code),
but I kept running into issues with my shortcuts, and have just resorted
to this.  This is obviously terrible, so I made sure that it was at
least somewhat kinda-sorta extensible should a way to improve it be
devised, which I hope there will be.

But hey, at least debug builds are possible now!