Added some checks
This wasn't a part of the project, but I decided to do it anyway.  One
of these checks doesn't work correctly, and I have no idea why.  I tried
rubber-ducky debugging it and it all looked completely correct, so I'm
going to assume that there's a bug in the library.