This commit is woefully incomplete.
So…this commit has done quite a bit.  First, it made the log
splashes work properly, which they didn't before, though that's
probably the least significant change.  Second, the beginning
phases of getting the API up have begun—at the moment, this is
just the ability to use the `-l` option to load a shared object
that defines a function `void gake_main(void)` and execute that
subroutine.  It also updated the manpage for `gake(6)`, and added a
new manpage:  `gake-api(7)`, an overview of the gake API.  Both of
these pages are currently tremendously innaccurate.  It also
made the makefile kinda better, and prepared it for the API, so it
will barf at you if you try to prepare a release build.  Finally,
it added a function `vlogmsg()` as a `va_list`-accepting
counterpart to `logmsg()`.  I think that's all.  It's quite a lot.