API finally works!
This commit took a while to make.  I could try to claim that this was
due to external influences…but no, this was literally just me struggling
with this.

Anyway, this commit has done quite a bit.  The biggest, most important
change is that the API _finally_ works—check the manpage to see how to
work it.  In addition, I've fixed the splashes (previously, the game
would fail to load one literally half the time), added a copyright
section to the manpages, modified the boilerplate comments, made note of
the usage of libgcrypt in the license (oops!), modified the makefile,
and removed the manpage on Gake's internals (since it was drastically
outdated).

Next steps:  Get a menu system up.  Then, at last, gameplay.